feat: Home page; solidify project structure
With this commit, the Home page is pretty much done. The only change
I expect is writing a more generic version of `genderLinks` and
`nameAndGenderInput` that also works on the upcoming Custom
Probabilities Character Generator page.

Helpers which make sense to put in their own modules have been, e.g.
`Anim.elm`, as it deals with some nitty-gritty internals of the
Animation library, and exports helpers which will definitely be used in
future pages. `Ctx`, which encapsulates its own state.

This introduces a few project structure improvements:
- Tighten up `Main.elm`, move `Route.elm` back into it
- Use color palettes and responsive size palettes
- Query the viewport on app startup to get the initial viewport size.
- Use the `Page` data type to group `title`, `init`, `update`, and `view` functions
- Working lens generator for erlandsona/elm-accessors V4
- Add some sensible exemptions to the elm-review config