fix: character URL rewriting logic
When the character page rewrites the URL from `/character/`
(RandomCharacter route) to `/character/name/gender`, `Main.elm`
is now prevented from rerouting (rerendering the page with the new route).
The reason we rewrite the URL is so that the user can copy the URL of
randomly generated characters.

Additionally, when going back home after the above flow, and `Home`
queries what the `Ctx.previousRoute` was, `Ctx` will say
`RandomCharacter`. This is achieved by Ctx tracking the last 2 previous
routes and treating the above flow as a special case.

Other minor changes:
- the character page now uses real data from the API now that we have
  elm-live setup
- `elm-live` now runs with --hot which tries to preserve Elm state upon
  recompiles
- The font is now loaded via `index.html` which allows us to use `block`
  `font-display` instead of `swap`.
  - Quick note on why `block` (read on about it for more context): FOIT
    is better UX than FOUT in game-like contexts. I would rather my app
    take longer to load than have an ugly FOUT.